move AuthTokenFile into gist.rb
