+++ lib/gist.rb
@@ -1,7 +1,6 @@
 require 'net/https'
 require 'cgi'
 require 'uri'
-require 'gist/auth_token_file'
 
 begin
   require 'json'
@@ -40,6 +39,27 @@ module Gist
     end
   end
   class ClipboardError < RuntimeError; include Error end
+
+  # helper module for authentication token actions
+  module AuthTokenFile
+    def self.filename
+      if ENV.key?(URL_ENV_NAME)
+        File.expand_path "~/.gist.#{ENV[URL_ENV_NAME].gsub(/[^a-z.]/, '')}"
+      else
+        File.expand_path "~/.gist"
+      end
+    end
+
+    def self.read
+      File.read(filename).chomp
+    end
+
+    def self.write(token)
+      File.open(filename, 'w', 0600) do |f|
+        f.write token
+      end
+    end
+  end
 
   # auth token for authentication
   #
